+++ private/iWorkImport/KNAnimationRegistry.h
@@ -20,7 +20,7 @@ __attribute__((visibility("hidden")))
 }
 
 + (id)localizedNameForUnsupportedAnimation:(id)arg1;
-+ (id)localizedCategoryNameForCategory:(int)arg1;
++ (id)localizedCategoryNameForCategory:(long long)arg1;
 + (id)categoryNameForPluginClass:(Class)arg1;
 + (id)animationsInBundle;
 + (id)instance;
