+++ private/iWorkImport/KNBuildAppear.h
@@ -22,7 +22,7 @@ __attribute__((visibility("hidden")))
 + (id)localizedMenuString:(long long)arg1;
 + (id)supportedTypes;
 + (id)animationFilter;
-+ (int)animationCategory;
++ (long long)animationCategory;
 + (id)animationName;
 - (id)animationsWithContext:(id)arg1;
